test
I realize that the year in the post is a grouping and not related to the post so I was overthinking trying to group the two up. (it would still be nice too, but too much for me to look into right now). I did move content_container to archive-single for now mostly to see what else it will affect.