update
I decided to remove content_container from html files. The _content.scss will be used to make custom version of certain elements related to content. The idea is to use the web inspector to locate the class and use that to modify a custom version. Also this is a test to see what happen if I try to override a class instead of modifying it.