Updated the default object handling in scene
In the JS version of the fishing widget, the scene objects are stored base on their top level class and their id is their array index. This is to allow better filtering by class and allow handling many objects over time since unused slots would be reused. id may be null or undefine though and the default case is a simple approach that may need to be overridden for other cases.