finish adding the notify_signal
Now scene object's destroy should notify the owning scene that it want to be destroyed. Other might be added depending if deem necessary, but notify_signal passes a id as a form of a message. This means the scene and scene object can be extended or edited to include additional cases without modifying the create object logic.