minor update
Added a bobber object that currently only move to where the mouse is clicked as well as toggles the visibility.

Also made the update cycle in the test_scene end if it was handled(such a clicking a stack fish will only removed one).

I delay the scene object destroy call to happen the next time it is updated, but this could be overridden to happen as it is called.