minor update
change how point's set works. It can handle a point if it is pass as the first parameter and will set its axie to a copy of the points. Not sure if I should loop and set instead since the axie array should not normally be ref elsewhere.

Also ran a test with finding the bobber and moving to it, but currently commented out since that will need a dedicated stated in the fish AI else they will be swarming the bobber when ever it is visiable.