minor update
added an equals function to the Point object so two different Points can be compare and will be consider equal if their axes are the same. Different dim points will treat unset axis as 0 by default.

Also uncomment the fish move to bobber test after adding a random chance to move to it. It still lacks the rest of the logic.