minor update
added velocity to the bobber movement so in the future the fish can fight back also allow the bobber to keep moving a little after input.

Also added a set case when adding or subtracting from a Point instance. This allow unset axis to be set if not 0 for the base Point class. Classes that extend Point could add more limits if needed.