minor update
Added the data to the fish, but may handle it differently in the future. (I may make a data access point and have the fish store it and use it to fecth fish data and use the fish id given on creation to fetch its data. This would allow it to be updated at runtime).

Also modify the fish data json. I replace the periods with tabs and added a name property since the key name should be id and name should be the version displayed(or key used in localizations)