minor update
Added displaying when the fish caught as well as adding fish weight. Also made it so fish weight effects how it is reel in, but it not tune to be too noticeable. Need to clean up that part more, but after I get around to adding reel in on holding mouse click (and after mobile support).