minor update
Added the reel in on mouse down logic. Now what left is clean up, adjusting data values, applying scaling base on distance(so the fish are smaller further away), and probably clearing the message over time or handling it a bit differently. Other stuff could be added, but the major goals have been reached for this sample.