minor update
Added scaling base on distance, but it a simple addition that need to be design better. It did require some parts of the canvas scene to be broken up to make it easier to modify.

Also move the game code out of the html and into it's own file. I may or may not move main back.

Note: I may try to make file names and dir lowercase overtime since I heard some systems do not like caps in paths, though I have not experience that.