minor update
Added a new js script that is meant to help with editing images. It also can cache the images as a url and I think it properly cleans up the urls. The plan is to allow it to provide tools for editing and caching images for later use. For the fishing project, I may use it as a way to apply shader like logic. I may or may not use it to modify the background.