minor update
Made the fish size depends on their weight. Also made the fish change when caught and relocate it outside the bounds (reusing the fish instead of destroying and recreating it or leaving it there to be catch again).

I still need to figure out best place to handle distance calculation. viewport seems ideal for rendering, but collision may need to define what is depth.