minor update
I did some clean up and move most style out of the html and into a css. Also made the js scripts behave as modules.

I also added a save system, but added a prompt to enable it or not. This will only appear if not clicking yes at least once. To disable it will need to edit the local storage and set 'save_locally' to false or remove it.