minor update
move the depth logic more to the viewport, but ended up calculating the 0-1 value back where it is used. Maybe I will add it to the scene interface since both cases need scene components (the bobber for adjusting position and the scene it self).
Note that the viewport has a world_position now.